Comp 401 Assignment1
This program reads the number of customers, each customer's first name, then last, then the number of items bought. It runs a for loop to then read the number each item bought and the name of that item. The output of the program is the number of each item bought, in alphabetical order of the names of each item on the menu.